# Ignore: 🔧 30x Performance Improvement in PreAuthorize AOP