--- a/pennyway-socket/build.gradle
+++ b/pennyway-socket/build.gradle
@@ -36,6 +36,7 @@ dependencies {
     implementation group: 'org.openapitools', name: 'jackson-databind-nullable', version: '0.2.6'
 
     implementation 'org.springframework.boot:spring-boot-starter-validation:3.2.3'
+    implementation "org.jetbrains.kotlin:kotlin-reflect:1.9.22"
 }
 
 tasks.withType(KotlinCompile) {