--- a/pennyway-socket/src/main/java/kr/co/pennyway/socket/common/util/PreAuthorizeSpELParser.kt
+++ b/pennyway-socket/src/main/java/kr/co/pennyway/socket/common/util/PreAuthorizeSpELParser.kt
@@ -71,7 +71,6 @@ object PreAuthorizeSpELParser {
             context.setBeanResolver(BeanFactoryResolver(applicationContext))
 
             method.parameters.forEachIndexed { index, parameter ->
-                println("parameter.name: ${parameter.name} -> args[index]: ${args[index]}")
                 setVariable(parameter.name, args[index])
             }
         }