--- a/pennyway-socket/src/test/resources/logback-test.xml
+++ b/pennyway-socket/src/test/resources/logback-test.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<configuration>
+    <include resource="org/springframework/boot/logging/logback/base.xml"/>
+    <logger name="org.springframework" level="INFO"/>
+    <include resource="kr.co.pennyway"/>
+    <logger name="kr.co.pennyway" level="DEBUG"/>
+</configuration>